[cookbooks] Don't manage SHMMAX and SHMALL
According to the postgresql documentation [1] a resonable value for
SHMMAX is now 1KB and that

> PostgreSQL requires a few bytes of System V shared memory (typically
> 48 bytes, on 64-bit platforms) for each copy of the server. On most
> modern operating systems, this amount can easily be
> allocated. However, if you are running many copies of the server, or
> if other applications are also using System V shared memory, it may
> be necessary to increase SHMMAX, the maximum size in bytes of a
> shared memory segment, or SHMALL, the total amount of System V
> shared memory system-wide. Note that SHMALL is measured in pages
> rather than bytes on many systems.

The previous values were required for Postgresql 9.2 which used
substantially more shared memory.

Removing the management of these system paramaters removes one of the
barriers to running the chef-server package directly in a container.

I've avoided doing any cleanup since this changed a core system
configuration which might also be needed by other software on the
system.

[1] https://www.postgresql.org/docs/9.6/static/kernel-resources.html

Signed-off-by: Steven Danna <steve@chef.io>