Enable FIPS mode in NGINX (and potentially other things)
This patch writes out some openssl config that tells the library
that we want FIPS mode by default. NGINX calls `OPENSSL_config(null)`,
which first checks the environment variable OPENSSL_CONF for a config
file, and then falls back to the system default. In this case, we
have written a system default for chef server with `fips_mode = yes`.

Signed-off-by: Jay Mundrawala <jdmundrawala@gmail.com>