[ctl-commands] Restart haproxy on upgrade in chef-backend mode
When configured to use Chef Backend, all postgresql requests are
routed through HAProxy. Thus, in order to ensure that we can access
postgresql during upgrades, we need to start haproxy rather than
postgresql.

NB: This assumes that the existing 15 second sleep is enough time to
allow haproxy to stabilize (i.e. mark non-leader nodes as DOWN). In
the reconfigure cookbook we do an active check that looks for the
steady-state rather than a sleep.

Signed-off-by: Steven Danna <steve@chef.io>