[ctl-commands] Orphaned auth_actor cleanup tool
Older versions of Chef Server produced orphaned auth_actor objects
easily. Some installations may have thousands of defunct entries
related to now-deleted (or never-created) clients. These defunct
entries can hurt performance and (in the worst case) make it
impossible to lis the membership of the group.

This command is based on a more straightfroward script created by
Josh Hudson:

https://gist.github.com/double-z/a062c9133719c2ac385a

It constructs a list of known actors (in the opscode_chef database)
and deletes any auth_actors in the bifrost database not contained in
the known actors list.

One issue with that script is that it requires the server to be
offline to run since otherwise newly-created clients might get
deleted.

This script is intended to run on-line. It uses a temporary new table
in bifrost to track newly created auth_actors. Any newly created
auth_actors are ignored when constructing batches of authz_id's to
ignore.

To avoid deleting clients whose creation spans the creation of the
tracking table and known actors list we use a one-time sleep.

Signed-off-by: Steven Danna <steve@chef.io>