[pedant] Wait for rabbitmq before every search test
The search index in eventually consistent but the test code treats it
as if it were read-after-write consistent.

To mitigate this we wait for the indexing queue to be empty before
executing a search test.

This is going to slow these tests down, but we are trying to stabilize
the tests first.

Signed-off-by: Steven Danna <steve@chef.io>