Remove use of skip_omnibus_build?
We can't use this helper right now as there is now PR (to read labels
off of) directly associated with the Automate Workflow change which is
driving the build out and testing of acceptance. We'll add this
functionality back in the near future when we've found a way to link
things together.

[ci skip]

Signed-off-by: Seth Chisamore <schisamo@chef.io>