Check for sufficient disk space before upgrade
The `pg_upgrade` tool copies all of the database files as part of the
upgrade.  While this could be avoided with the `--link` flag, the
`--link` flag would also prevent falling back to the old pg cluster in
case of a problem.

To avoid upgrade failures, we check that the disk for the new data
directory has enough free space for another copy of the database plus
some headroom to ensure we have some room to operate after the
upgrade.

Available disk space is currently queried via an FFI call into libc's
statvfs function. The downside of the FFI approach is making sure our
representation of the statvfs struct is correct on all of our
supported platforms. A more straightforward approach might be to parse
`df` output; however, we've gone with the FFI call for now to avoid
some edge cases in parsing `df` output on versions of `df` without an
--output flag.

Signed-off-by: Steven Danna <steve@chef.io>