[cookbooks] Make pg_upgrade timeout configurable
The default timeout for an execute resource is 3600 seconds. For large
Chef databases (such as those with a lot of reporting data), this may
not be enough time for pg_upgrade to succeed.

This makes the timeout configurable and sets the default to 7200 which
will hopefully be high enough for most production Chef Servers.

Signed-off-by: Steven Danna <steve@chef.io>