Mitigate Host Header cache-poisoning attacks
Previously, a request such as:

    curl -H Host:evil.com http://chef-server.dev

would create a response where the Location was set to the user
provided Host header.

    Location: https://evil.com/

If this 301 was erroneously cached, it may then be served to another
user, whose traffic might be redirected to an attacker-controlled
server.

This PR changes the nginx configuration so that we use the server_name
rather than the user-provided Host header when generating redirects.

Signed-off-by: Steven Danna <steve@chef.io>