Check for chef-manage package
So it turns out that if we set these headers from the API, and the user
also has chef-manage installed, then they get duplicate headers.  Adding
a check to see if the chef-manage package is installed before adding the
headers.

There are potential edge cases here if a user has the manage package
installed, but not configured/running.  In those cases, the headers
would be missing.  The path forward would be to completely purge the
chef-manage package from the system and reconfigure.

Signed-off-by: Nolan Davidson <ndavidson@chef.io>