Disable Welcome Page
This allows a user to disable chef server's welcome page.
When the welcome page is disabled, a 404 will be returned when a request
is made that does not match any needed pattern.

To disable the welcome page add the below line to the chef-server.rb
file.

`nginx['show_welcome_page'] = false`

The change is only a cookbook change that modifies nginx's config file

This would close this issue:
https://github.com/chef/chef-server/issues/972

Signed-off-by: Lance Finfrock <lfinfrock@chef.io>