[partybus] Store migration-level in postgresql database
The migration-level was previously stored in a flat-file on disk.
This file then had to be specifically accounted at multiple points in
a Chef Server installations life cycle.

By placing the migration-state in the database, we hope to decrease
problems related to incorrectly synced migration-level files.

Both the /infer/ and /upgrade/ commands will import existing
migration-level files into the database if the database is not already
initialized.

TODO:

- [ ] For now I've treated PostgreSQL as a key-value store and didn't
  change what metadata we store. We should decide now if we think we
  want to do something fancier.

Signed-off-by: Steven Danna <steve@chef.io>