[oc-chef-pedant] Normalize chef server uri
When rendering the chef server URI in pedant.rb config files, we want
to include ports (for example when the chef server is behind a
proxy). That creates a problem when we use the standard port, as the
URI gets normalized many places in the chain from pedant to erchef; in
particular redundant port specifications (e.g 443 for https) are
stripped out. However the specs will retain the port, causing test
failures. We normalize the URI here to make sure that the specs we
check against conform to that requirement.

Signed-off-by: Mark Anderson <mark@chef.io>