removing forcing of the doc formatter
the output is so long that it blows out travis's line limits and the way
it gets forced here is difficult to override.

it can go into .rspec files in projects to bring it back if people
really like it by default, but its very wrong to have it here.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>