[oc-id] Commit assets to git and remove nodejs
Historically we built assets in the omnibus pipeline, but that has
proven difficult because nodejs no longer supports all of our
platforms.

Instead we are committing assets to git. This isn't ideal, as it makes
it harder alter assets in the dev-vm; they are now synchronized from
the host system. However, we don't update oc-id often, which mitigates
this somewhat.

This includes several changes:

1) Remove nodejs and the asset precompile stages from the build of
   oc-id. This incidentally removes nodejs from our manifest. We
   delete it from the final build, but some security scans have
   flagged it anyways.

2) Commit the used assets in src/oc-id/public/id/assets to git.

3) Modify dev-vm to sync those assets to the vm

4) Update the oc-id README.md with instructions on building & updating assets.

Signed-off-by: Mark Anderson <mark@chef.io>