Pin ci-studio-common to pre-1.0 branch.
ci-studio-common 1.0 was released today with lots of breaking changes
which are not fully compatible with travis containerized builds. This
pins the branch to a pre-1.0 branch rather than switching to slower
GCE dedicated builders to consume the new release. Longer term we
intend to switch over to the buildkite driven infrastructure and then
can update this to use the latest release.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>