[chef-server-ctl] Make chef-server-ctl configurable
This PR adds the ability to inject Chef Server configuration items
required by chef-server-ctl subcommands. A Config class has been added
that provides access to this configuration globally.

The short-term goal is to make it straightforward to get more of the
chef-server-ctl commands working from Habitat packages.  We now have 2
sets of Habitat packages:

- the docker-focused Habitat packages in this repository
- the A2-focused Habitat packages in the A2 repository

The A2 packages consume the packages in this repository.

Previously, the server-admins, key-rotation, reindex, bifrost-cleanup,
filtered-dump, and password commands were unlikely to work with either
set of packages. Now, these commands can work in our Habitat packages
if the wrappers set the appropriate configuration.

The long-term goal of this is two-fold. First, we want to make
chef-server-ctl are more standalone application. The first step of
this was the work that moved the command into its own gem. In addition
to supporting Habitat, making these dependencies configurable means we
can do things like:

- Inject stub versions of commands to test failure conditions Point

- chef-sever-ctl and chef-zero or similar API stubs for testing

- Decouple sub-command logic from the particulars of omnibus
  packaging, paving the way for future packaging changes.

The second long-term goal is to make the Habitat packages in Chef
Server more generically consumable for downstream consumers such as
A2.

* Ugh, why environment variables?

In the long-run I think we might want a custom configuration file
based on mixlib-config or similar. However, environment variables
have the following benefits:

- Easy to set from Habitat wrappers
- Easy to set at the command line for testing
- Doesn't require immediately fixing all of chef-server-ctl's command
  line argument processing bugs
- Avoids pre-baking a config-file format until we know what we need.

I think we should consider these environment variables UNSUPPORTED at
the moment and thus developers are free to change them

* Should we get rid of CHEF_SECRETS_DATA in favor of this?

No, at least not yet. Getting rid of CHEF_SECRETS_DATA in chef-server
itself will require being very careful to not add any credentials to
any new on-disk files.

* Why didn't you fix all of the commands?

I've limited this initial change to the commands supported in the A2
Habitat packages. This also has the benefit of not touching some of
the more complex commands just yet. The un-modified commands include:

  - reconfigure
  - oc-id commands
  - secrets rotation commands
  - All runit-based commands
  - HA commands
  - upgrade/chef-11-upgrade
  - gather-logs

Fixing all of chef-server-ctl will be a lengthy process since we have
a number of environments to test it in.

Signed-off-by: Steven Danna <steve@chef.io>