Postgres listens on loopbacks, not localhost
We previously set postgresql to use `listen_addresses = 'localhost'`
for standalone installs. But that can be problematic, depending on the
/etc/hosts file; on some platforms that resolves to the ipv6 loopback,
and postgres seems to take this as ’listen only to ipv6 loopback (::1)
’.  A lot of the services use hardcoded ivp4 loopback to connect, and
fail.

This should not affect tiered or HA systems, since in those cases we
listen on a specific IP address, not localhost.

Signed-off-by: Mark Anderson <mark@chef.io>