[WIP] Update cookbooks to use runit custom resource
The runit cookbook was recently updated to use an init.d script rather
than symlinking the sv binary to /etc/init.d. The template used by
runit hard-codes the PATH variable which breaks the omnibus path to the
`erl` binary, and causes errors to be displayed when starting the
service via chef-server-ctl. In effort to resolve this issue, I have
been working on some changes to the upstream enterprise-chef-common
cookbook in conjunction with these changes.

Unfortunately, despite my efforts this still doesn't quite work as the
I cannot seem to override the template source using edit_resource as
the runit_service custom resource seems to make the scope of the
`template['/opt/omnibus/init/rabbitmq']` unavailable, or is not
compiled in time for `edit_resource` to function properly.

I think this work may still have merit, so I am committing it here as
WIP so I can look at it later.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>