Remove applications that depend on wx
The et, debugger, gs and observer applications depend on the wx
library, which in turn depends on X11's wxWidgets library. We don't
want to include that in omnibus, so we must not use these apps in our
releases.

We were able to include these because despite being built without wx,
we were shipping those applications as part of erlang.  This was an
artifact of us using the source tarball from the downloads site, which
includes prebuilt beam files to save time, including those that
require wx. Moving to git based erlang builds starts at an
earlier stage that lacks those prebuilt files.

Signed-off-by: Mark Anderson <mark@chef.io>