Trigger nightly build with Expeditor
Now that we removed the nightly trigger from the Jenkins pipeline (per
chef-cookbooks/opscode-ci#936) we need to update the Expeditor config to
begin triggering a nightly build of the project. This build also expires
the Omnibus cache so we can be sure Chef Server builds correctly even if
some transitive deps have been udpated.

Signed-off-by: Seth Chisamore <schisamo@chef.io>