[chef-server] Remove do_end function from chef-server-ctl hab plan
As far as I can see, this was probably copy-and-pasted from another
plan. In this plan we do not set _clean_env anywhere so we do not need
to clean it.

In reality, I need to rebuild this package in an attempt to get out of
a dependency mess downstream.

Signed-off-by: Steven Danna <steve@chef.io>