Rework the erlang ct test harnesses to be a bit more maintainable.
The ct harnesses have some behaviours that made life a little difficult.

1) They did a lot of database setup/teardown in erlang, which was hard
for non erlang people to maintain, especially in the context of CI
systems.

2) In some cases they required the ability to create and destroy a database locally

3) It proved to be difficult to get it working in the new buildkite docker container.

The new behaviour is meant to work with pg_virtualenv if desired, but
also work if the user has access to postgres directly.

One thing to note is that the ct tests are fragile, in that if setup/teardown fails the tests themselves are silently skipped.

Signed-off-by: Mark Anderson <mark@chef.io>