render profiles and data-collector upstreams correctly
when setting the configuration value `insecure_addon_compat`, the
chef-server library code does not render the secrets from the
secrets helper into the PrivateChef hash, which results specifically
in `data_collector['token']` not showing up there. when configured
to proxy compliance profiles and the data-collector endpoint, however,
the nginx recipe was relying on this field being present for the
upstreams to be rendered into the config. this resulted in us never
rendering the correct upstreams in `insecure_addon_compat` mode.

this change makes it so that the profiles and data-collector upstreams
will always be rendered if their root_url is configured. if, by
chance, the user doesn't have the `data_collector token` secret set,
we'll now see a 401 response code and an error message instead of a
confusing and misleading 404.

Signed-off-by: Stephen Delano <stephen@chef.io>