Alter the delivery build recipe to accept the chef license.
Chef client has been updated to v15 on our delivery builders, and now
needs to have the license accepted. We have a somewhat heterogeneous
environment, and this recipe is actually running on chef client 13,
while the parallel_remote_execute resource is running a chef-client
command on a build node running chef 15. Using environment variable
form because I'm not sure if we've globally updated our system.

Signed-off-by: Mark Anderson <mark@chef.io>