[private-chef-cookbooks] Kill rabbitmq if sv stop times out.
As described in issue #1690, the first run of `chef-server-ctl
reconfigure` fails with a timeout at the very end of run, when
executing delayed restarts:

    STDOUT: timeout: run: /opt/opscode/service/rabbitmq: (pid 7674) 68s

Successive runs of reconfigure work fine.

This appears to be a problem buried deep inside the rabbitmq restart
logic; we weren't able to make much progress with the root cause.

The strategy of stopping then starting doesn't seem to help our
restart problem. The stop times out, breaking the reconfigure.

We can add a execute block with default action nothing that
aggressively shuts down rabbitmq (normal shutdown followed by pkill if
that doesn't work), then waits, then starts up rabbitmq again. Instead
of sending a restart action (or the stop/start here) we send the
execute action to this block.

See also:

https://wiki.openstack.org/wiki/OpsGuide-Maintenance-RabbitMQ#RabbitMQ_service_hangs

Note: it's worth evaluating the consequences of not restarting it at
all; mainly the restart is to enable the rabbitmq management plugin,
which if might also be broken by the same issue that's making the
restart timeout.

Signed-off-by: Mark Anderson <mark@chef.io>