Implement #1703 exclude data_collector from _status endpoint
Adds a configuration option to allow the data_collector health check to be excluded from the overall health reported by the _status endpoint.

Add `data_collector['health_check'] = false` to `/etc/opscode/chef-server.rb` and reconfigure to disable data_collector health checking.

When using the _status endpoint to control which frontends are included in the load balancer pool, perfectly good frontends are removed from the pool if the data_collector endpoint (Automate 1 or 2) is down.

This causes an unnecessary Chef Server outage as all the frontends are perfectly able to service requests.

Reported by 2 customers.

Signed-off-by: Richard Nixon <rnixon@chef.io>