enable oc-id to receive a dev config
This updates the chef-server.rb template to write-out an oc-id
configuration if one is present. The config will be present if it is
added to either dev/defaults.yml or the "personal" dev/config.yml in
the form of:

```
vm:
  ...
  oc_id:
    administrators:
      - admin
    applications:
      # provide an app name and its redirect_uri
      supermarket: http://localhost:3000/auth/chef_oauth2/callback
```

The Vagrantfile has been updated to merge the oc-id YAML data into the
JSON used for provisioning and ultimately by the chef-server.rb
template.

Signed-off-by: Robb Kidd <robb@thekidds.org>