turn off audio on the dev VM
No need to ruin my tunes.

(If audio is enabled, VirtualBox will attach a VM to the host's audio
system at boot. This includes both input and output. If you're listening
to music on Bluetooth headphones, you will notice the audio quality go
down as the device switches to bi-directional mode to be able to send
audio from the microphone. The dev VM doesn't need sound, so let's turn
it off and making the development experience sound better.)

Signed-off-by: Robb Kidd <robb@thekidds.org>