fix SIGTERM breakage for erlang 20.x
Changes to SIGTERM handling in Erlang 19.x caused runit failures in
Chef Server. This was originally fixed when we upgraded to Erlang 19
by passing a startup flag to the Erlang VM on bootup.  But the startup
flag was removed in Erlang 20, so another fix had to be employed.

This change effectively reverts Erlang's SIGTERM handling to previous
(pre-19) behavior.

REF: https://github.com/erlang/otp/pull/1315

Signed-off-by: Lincoln Baker <lbaker@chef.io>