configure build, test, verify pipelines for erlang 19
The version of erlang defined in the bk_install.sh is the version used for testing in the verify pipeline.  If the version required is not available in the docker image, it can be installed and used in bk_install.sh.  The version in .expeditor/verify is the one we want to use for testing, but there is no guarantee that it will be used, unless it is installed in the base docker image.  Hence we also print out the env variable in the script to affirm the actual version of erlang being used for testing.

changes:

fix buildkite erlang version
delete unused code referencing erlang version
log the version of erlang and the env currently being used for testing.
remove reference to ERLANG_VERSION=19.3 in verify.pipeline.yml
remove erlang install step from bk_install.sh

Signed-off-by: Lincoln Baker <lbaker@chef.io>