ci: disable Microsoft apt mirror, fail fast in bk_install.sh
The verify buildkite pipeline started failing with the following error
during apt-get update:

    E: Failed to fetch
    https://packages.microsoft.com/ubuntu/18.04/prod/dists/bionic/main/binary-amd64/Packages.gz
    File has unexpected size (93512 != 165979). Mirror sync in progress?
    [IP: 13.74.252.37 443]
       Hashes of expected file:
        - Filesize:165979 [weak]
        - SHA256:179eb71f2afb4a72bf5b11180b4d4c9ccf1644076dd75f5a7bbf880ecefafbba
        - SHA1:381a8321619083a4063fa8381bf3aa12a2dac5a3 [weak]
        - MD5Sum:54c730dd6a33c612b2ae3c23fe0cfcb7 [weak]
       Release file created at: Thu, 12 Dec 2019 19:59:19 +0000
    E: Some index files failed to download. They have been ignored, or
    old ones used instead.

In previous cases of this error, it has taken many hours for the
upstream repository maintainers to fix the problem. For example, see:

   https://github.com/dotnet/core/issues/3221

Since we do not depend on software in this repository for our tests,
we've temporarily removed the source.

Further, we've made the bk_install.sh script exit immediately on error
and added some logging to make it a bit easier to debug in the future.

Co-authored-by: Prajakta Purohit <prajakta@chef.io>
Signed-off-by: Steven Danna <steve@chef.io>