Replace self-signed certs in external postgresql scenario
The external PostgreSQL scenario now replaces the private key and
self-signed certificate that are generated by the PostgreSQL installer
with a real certificate from Vault if the scenario is being executed by
Buildkite.

The scenario has also been modified to begin with SSL disabled and a
smoke test followed by enabling the enforcement of SSL from both the
PostgreSQL config AND the Chef Infra Server config before executing the
remaining tests.

Signed-off-by: Christopher A. Snapp <csnapp@chef.io>