Update listen_port for newer nginx.conf syntax.
The old listen_port helper was designed for older nginx.conf syntax (when
"ssl on;" was necessary for HTTPS.) "ssl on;" has been removed from the
template for chef_http(s)_lb.conf, and the listen_port helper has been
updated to append the ssl option (e.g. "listen 443 ssl;") when "https" is
passed as proto.

Signed-off-by: Thomas J. Gallen <kaori.hinata@gmail.com>