wip remove accessing request payload in effort to cure 500 possibly caused by webmachine
the 500 i've been hunting is possibly in webmachine.  if a payload
(request body) is ever accecced via wrq:req_body, it can never be
accessed later in the same request by wrq:stream_req_body. however later
on (not in this module) the payload is apparently accessed by
wrq:stream_req_body, thus the possible source of the 500.

Signed-off-by: Lincoln Baker <lbaker@chef.io>