[cookbooks/elasticsearch] Bind to correct address and fix other warnings
We were using the wrong attribute (`ip_address` instead of `listen`)
when setting the backend defaults, leading to ES only listening on
localhost on the backend.

After fixing this, I also found that we were failing bootstrap checks
that ES enforces once you are listening on a public interface.

Finally, this fixes two warning related to bad configuration we seemed
to be passing (there are more warnings remaining).

Signed-off-by: Steven Danna <steve@chef.io>