[terraform] Fix terraform syntax problem
I don't know how this is working for other people, but terraform
complains about these lines previously using a non-boolean expression
in a ternary.

My guess is that it is some terraform version difference, but this
test suite as used up enough of my time today so I haven't dug in
further.

Signed-off-by: Steven Danna <steve@chef.io>
Signed-off-by: Christopher A. Snapp <csnapp@chef.io>