reimplement expiration windows.
The integration testing pipeline exposed a failure in chef backend.
Subsequent investigation revealed the failure was caused by a feature
we will describe as 'expiration windows,' the purpose of which was to
create cache-friendly URLs.

"expiry windows" cannot be done the same way in sigv4 as sigv2.
Sigv4 X-Amz-Expires operates on a seconds until expiration basis, as
opposed to an explicit expiration timestamp in sigv2. sigv4 timestamps
would also need to account for the x-amz-date value.

The caching use case requires stable signatures; manipulating either the
X-Amz-Expires or X-Amz-Date will not automatically produce stable signatures.

for reference:
https://blog.chef.io/cookbook-caching/
https://github.com/chef/mini_s3/blob/master/src/mini_s3.erl (expiry_time/2, etc)
https://docs.aws.amazon.com/general/latest/gr/sigv4-date-handling.html

Signed-off-by: Lincoln Baker <lbaker@chef.io>