Initial chef-server-ctl wordmark replacements #1949
This is an initial pass of replacements for issue #1949: Make Chef
Server distributable under a configurable name - wordmarks in outputs.

This pas focuses on `chef-server-ctl` and replaces:

 * `chef` with `#{ChefConfig::Dist::SHORT}`
 * `Chef Server` with `#{Chef::Dist::SERVER_PRODUCT}`

Multiple passes will likely be required; this is a small chunk meant
to be easy to review and test.

This is a combination of 3 commits:

 * f26fdb98915515fb10a86e20f2f8a25fef8e4170
 * 03c8686637264378eefafd7b22f5c7a363c2d14f
 * 7fae6f49a2c6b1e98071aa5acbb3fe37b26af2b9

Signed-off-by: Josh Gitlin <jgitlin@pinnacle21.com>