Use runit custom control to workaround Erlang signal handling
Starting in OTP 19, SIGTERM calls init:stop() to perform a graceful
shutdown. However, it appears that this behavior does not work for
us. Attempts to restart the service with this behavior in place
results in a timeout from runit:

    ---- Begin output of /opt/opscode/embedded/bin/sv restart /opt/opscode/service/opscode-erchef ----
    STDOUT: timeout: run: /opt/opscode/service/opscode-erchef: (pid 22355) 66s, got TERM
    STDERR:
    ---- End output of /opt/opscode/embedded/bin/sv restart /opt/opscode/service/opscode-erchef ----
    Ran /opt/opscode/embedded/bin/sv restart /opt/opscode/service/opscode-erchef returned 1

In b6d9d74e0611b6d08d268a0c8ba6202233ae30d1 we avoided this new
behavior by setting ERL_ZZ_SIGTERM_KILL="true" in the environment,
instructing Erlang to keep the default signal disposition for SIGTERM.

When we upgraded to OTP 20, this workaround was removed and replaced
with a new workaround in
78c6a54a54d85d74c3a1a6be493f676eb1209711. Here, we used

    os:set_signal(sigterm, default)

at the beginning of our application's start function to reset the
signal's disposition to its default disposition (terminate the
process).

This workaround, however, appears to have a problem. If two runit
restarts are issued in close succession (as may happen during a
reconfigure), it is possible that the erlang process receives the
second SIGTERM before the application has had a chance to reset the
disposition, resulting in a runit timeout for the second restart.

The _REAL_ fix is to investigate why the new behavior doesn't work for
us and fix that. chef/chef-server#2106 has been opened to track this
work.

For now, this change uses runit's custom controls facility to send a
SIGINT rather than a SIGTERM. The erlang VM leaves SIGINT set to its
default disposition (terminate the process) and thus we should avoid
the race condition or the problematic init:stop behavior.

Signed-off-by: Steven Danna <steve@chef.io>