cookbooks: remove deprecrated_solr_indexing
In f47434bf6046d9c1bb02e62d2a5251f2752ee48b we added
deprecated_solr_indexing. This option is no longer supported. While
this didn't ship in a stable release, it did ship in a development
release that at least a few customers have used and thus we are
keeping a validation check to fail if it is present.

As part of this change I've added a new IndexingPreflightValidator
where we can move values from the old SolrPreflightValidator.

Signed-off-by: Steven Danna <steve@chef.io>