Remove opscode-expander
Opscode Expander was a Ruby service responsible for expanding node
objects into the schema expected by our search index.

This commit removes the service as we now do this expansion directly
inside Erchef. The new indexing pipeline allows us to remove the
operational complexity of RabbitMQ for our users. The new indexing
pipeline is not as battle-tested as Expander, but we hope that over
time it can reach the same levels of reliability.

Expander was an excellent service. It performed well in our Hosted
Chef environment and for our many users. It had remarkably few bugs in
its history and is a good example of the skill of its authors.

Signed-off-by: Prajakta Purohit <prajakta@chef.io>