erchef: new ibrowse to address spurious 500s
This pulls in a new ibrowse branch that has
chef/ibrowse@1df08a6aca30614f69126451957af6a3d0aeb7b2

The new commit avoids extra calls to ssl:send() with empty request
bodies.  We think that these extra ssl:send calls were causing
spurious 500s.

Based on some debugging of the problem, we noted that:

- mini_s3 uses HTTP/1.0 when making HEAD requests. Code comment
  indicates this is to avoid a bug in ibrowse related to chunked
  encoding.

- ibrowse always attempts to send a body, even when that body is
  empty. erchef is configured to talk to erchef via nginx.

Given this and some other debugging, we believe the following is
happening occasionally:

- ibrowse sends the start of the head request, including the headers.
  nginx gets this, sees content-length is zero so it can process the
  request without reading the non-existent body.

- nginx gets the response and returns it, and closes the
  connection. The connection is closed because mini_s3 used HTTP/1.0
  which doesn't support keepalive connections.

- ibrowse tries to call ssl:send(Sock, []) to send the empty body but
  gets {error, closed}.

- ibrowse returns an error for what would have been a successful
  request.

Avoiding the empty `ssl:send` resolves the issue in our reproduction
case.

We may also update mini_s3 to avoid using HTTP/1.0, since the
consensus of previous maintainers is that the bug it is working around
is likely no longer an issue. However, changing the HTTP version will
also change how the ibrowse pool behaves and thus we want to do that
as a separate change.

Signed-off-by: Steven Danna <steve@chef.io>