Wait and retry index_exists? check while Elasticsearch starts and stabilizes
We check if the elasticsearch index exists before creating it.
But Elasticsearch returns 404 for a short duration during startup before it stabilizes and can return a 200.
As a result, we incorrectly try to recreate an index when it actually exists in the system.
Re-creating an existing index is a 400 "Bad Request" error.

Signed-off-by: Prajakta Purohit <prajakta@chef.io>