Address @ssd's feedback on encode_sign_error_response (see PR)
stevendanna:

It's hard to see in this diff, but I think we need to return this
encode_sign_error_response in one more case which is where the
IncomingSig != CalculatedSig

@lbakerchef:

I checked what's on master regarding this
(https://github.com/chef/chef-server/blob/master/src/bookshelf/src/bksw_sec.erl).
From what I can tell, it seems to handle it differently depending
on whether it's a presigned URL or an authorization header.
For presigned URLs it seems to do it the way I do it above -
either send an encode_sign_error_response, or
encode_access_denied_error_response. For authorization headers,
it doesn't ever send an access denied, it just does an encode_sign_error_response.
Above, it seems I have factored all of this into a single function
that just does it one way - the presigned URL way.

Would you like me to change this to match the old behavior?

Signed-off-by: Lincoln Baker <lbaker@chef.io>