Fix pedant for ruby-3
URI::encode is gone forever, Addressable::URI.encode is a better
replacement.

This means we need to take an explicit dep on addressable, but
addressable was always in the bundle here anyway.

Since addressable itself is a dep nightmare and the API we're
using has been stable for at least 5 years, don't include any
pin here and let it float to avoid conflicts.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>