[chef-server-ctl] Need a runtime dep on chef in gemspec
Why:
It is required by some of the plugins in chef-server-ctl like
backup.rb (chef/mixin/deep-merge)
reindex.rb(chef/config)

Why did it work before:
In the past since there was no pin on the version of chef, this did not
matter.

Why adding the pin only to Gemfile did not work:
We use the older version of appbundler (takes 2 vars)
This ignores the deps pinned in Gemfile but not added to the gemspec.

To test:
In the appbundle (/usr/bin/chef-server-ctl) on local install there is was no pin on chef.
Manually adding this pin allows the reconfigure to succeed.

Signed-off-by: Prajakta Purohit <prajakta@chef.io>