Add Content-Security-Policy and X-Frame-Options headers to our static content
This came up in a security audit. The user browseable page that just points user to use knife or access our API does not set Content-Security-Policy and doesn't set `X-Frame-Options DENY` if manage is installed. This static content should be pretty locked down. I've tested this on an installation and it doesn't impact the page loading, but improved our scan scores.

Signed-off-by: Tim Smith <tsmith@chef.io>