Switch veil and omniauth-chef to gems not git sources
We have working pipelines for both of these projects now so we can use
git releases. This reduces the amount of spam from dependabot where
every minor bump of those projects gets pulled in here and it also
speeds up the installs since we don't have to clone both of these repos
during the build.

Signed-off-by: Tim Smith <tsmith@chef.io>