Construct PostgreSQL version object using a regex.
The PostgreSQL version comes in as string, for example:
13.3 (Ubuntu 13.3-1.pgdg16.04+1)

Naturally, the version varies and could be anything.

For our purposes, we want to convert the above example to:
13.3

Currently we do this via:
v = Gem::Version.new(r[0]['server_version'].to_f)

This works as long as PostgreSQL does x.y versions but it will
fail in the future if they move to x.y.z.

To future-proof this, I propose conversion via the following:
Gem::Version.new /^([0-9\.]+)/.match(r[0]['server_version'])[0]

Signed-off-by: Lincoln Baker <lbaker@chef.io>