WIP Don't run PostgreSQL's pg_upgrade for minor version upgrade
adhocs and umbrellas were working without this change, so it may not be
necessary.  PostgreSQL's pg_upgrade is probably smart enough to not do
things it shouldn't in case of a minor version upgrade.  however,
@prajakta did mention not running pg_upgrade for minor version upgrades,
so maybe this is necessary to avoid having to use *-bin files for
every minor version upgrade, or maybe it's necessary for some other
reason?

Signed-off-by: Lincoln Baker <lbaker@chef.io>