Use chef_cbv_cache to store frequently calculated cookbook responses
This makes use of the new `chef_cbv_cache` to minimize the amount of
times overall that a given formatted cookbook response is re-created,
because this operation is very expensive when the list of cookbook files
is long.  Since ~OTP 20 the impact of these CPU-intensive responses has significant
impact on the overall health and responsiveness of the system under
load.

By significantly reducing the frequency that we build a response, we reduce the
overall VM load; and the relatively infrequent number of times a new
response has to be built has much less impact on the VM.

Some numbers, using sample org data that tends towards generating large
responses to cookbook_version endpoint, because of the total number of
cookbook files:

7k clients/15min interval - cache disabled - CPU 1800-2000%, LA 25. cookbook_versions rough average time range 2-5s
7k clients/15min interval - cache enabled - CPU 150-200%, LA 5-6. cookbook_versions rough average time range 0.2-1.1s (edited)
40k clients/15min interval - cache enabled - CPU 1000-1200%, LA 19 . CBV time range: 0.2-3s with most on the < 0.5 end.
40k clients/15min interval - cache disabled - server fails within a minute

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>