Don't even start the cache if it's disabled
Based on code review feedback, this adds an up-front call to envy:get to
determine if the cache is enabled before doing breaker check or sending
the message to the cache.  The environment lookup calls are
lighter-weight than round-tripping two messages for each call for a
disabled service.

oc_chef_wm_sup will also no longer start chef_cbv_cache if it is disabled.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>